move more stuff to mesh client
